newly added css bootstrap

+++ database/migrations/2025_09_01_113707_create_heads_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->string('name',50);
             $table->string('surname',50);
             $table->date('birthdate');
-            $table->integer('mobile');
+            $table->bigInteger('mobile');
             $table->text('address');
             $table->string('state');
             $table->string('city');
